+++ history.md
@@ -6,6 +6,7 @@
 - 内部函数更新为$开头
 - 优化相关类，继承优化，输出优化
 - setLocalData中对JSON.stringify进行错误catch
+- require扩展delete/put方法
 
 ### 2.2.10
 - formatTree错误修正